requirements.txt
fastapi
uvicorn
openai
chromadb
pydantic
python-dotenv